# Conflicts:
#	src/base_station_base.h
#	src/newgrf.cpp
#	src/newgrf_class.h
#	src/newgrf_class_func.h
#	src/newgrf_commons.cpp
#	src/newgrf_commons.h
#	src/newgrf_roadstop.cpp
#	src/newgrf_roadstop.h
#	src/newgrf_station.cpp
#	src/saveload/newgrf_sl.cpp
#	src/saveload/saveload.h
#	src/saveload/station_sl.cpp

+++ src/base_station_base.h
@@ -24,7 +24,7 @@ extern StationPool _station_pool;
 struct StationSpecList {
 	const StationSpec *spec;
 	uint32 grfid;      ///< GRF ID of this custom station
-	uint8  localidx;   ///< Station ID within GRF of station
+	uint16 localidx;   ///< Station ID within GRF of station
 };
 
 struct RoadStopSpecList {
